apps/server/pyproject.toml
[tool.poetry]
name = "l3agi"
version = "0.1.0"
description = ""
authors = ["Giga Chkhikvadze <giga.chkhikvadze@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.13.3"
langchain = "^0.1.11"
fastapi = { version = "0.99.1", extras = ["all"] }
requests = "2.31.0"
pydantic = "1.10.11"
psycopg = "3.1.9"
jupyter-client = "8.3.0"
snakemq = "1.6"
pandas = "2.0.3"
matplotlib = "3.7.2"
pillow = "*"
gql = "*"
azure-messaging-webpubsubservice = "1.0.1"
azure-identity = "1.13.0"
faiss-cpu = "1.7.4"
google-search-results = "2.4.2"
uuid = "1.30"
sentry-sdk = { version = "1.26.0", extras = ["fastapi"] }
dateparser = "1.1.8"
psycopg2 = "*"
psycopg2-binary = "*"
arxiv = "*"
duckduckgo-search = "*"
wikipedia = "*"
langchain-experimental = "0.0.17"
zep-python = "*"
llama-index = "^0.10.16"
sqlalchemy = "*"
fastapi-sqlalchemy = "*"
alembic = "*"
nltk = "*"
ipython = "*"
tiktoken = "*"
requests-toolbelt = "*"
oauth2client = "*"
strawberry-graphql = { version = "*", extras = ["fastapi"] }
pymysql = "*"
youtube-search = "*"
pyowm = "*"
bs4 = "*"
llama-hub = "*"
twilio = "*"
boto3 = "*"
llama-cpp-python = "*"
pypdf = "*"
torch = "*"
transformers = "*"
sentencepiece = "*"
docx2txt = "*"
python-pptx = "*"
ebooklib = "*"
html2text = "*"
nbconvert = "*"
protobuf = "*"
s3fs = "*"
pinecone-client = "*"
weaviate-client = "*"
black = "*"
ruff = "*"
isort = "*"
huggingface-hub = "*"
replicate = "*"
sendgrid = "*"
fastapi-jwt-auth = "^0.5.0"
arrow = "^1.3.0"
deepgram-sdk = "^2.12.0"
langchain-community = "^0.0.25"
langchainhub = "^0.1.15"
langchain-openai = "^0.0.8"
llama-index-embeddings-langchain = "^0.1.2"
llama-index-llms-langchain = "^0.1.3"
llama-index-vector-stores-pinecone = "^0.1.4"
llama-index-vector-stores-weaviate = "^0.1.3"
llama-index-vector-stores-zep = "^0.1.2"
llama-index-llms-openai = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
